Abstraction Begins
- Bulk Sell Feature
- Single Flip Cards
- Abstracted ASCII elements so I dont set fire to the AI Generators
- Removed weird lines
